refactor: extract 3 components from TaskDashboardView (1384→640 lines)
Extract OrchestratorStatusCard, QuickPresetsCard, and RecentJobsTable
into dedicated components under components/tasks/. Moves ~750 lines
of template and scoped CSS into self-contained SFCs with clear prop
contracts. The orchestrator card owns its own store access, uptime
clock, and health state derivation.
