+++ apps/desktop/src/views/tasks/TaskDashboardView.vue
@@ -1,7 +1,10 @@
 <script setup lang="ts">
-import { ErrorState, formatDate, LoadingSpinner, SearchInput, StatCard } from "@tracepilot/ui";
-import { computed, onMounted, onUnmounted, ref } from "vue";
+import { ErrorState, LoadingSpinner, SearchInput, StatCard } from "@tracepilot/ui";
+import { computed, onMounted, ref } from "vue";
 import { useRouter } from "vue-router";
+import OrchestratorStatusCard from "@/components/tasks/OrchestratorStatusCard.vue";
+import QuickPresetsCard from "@/components/tasks/QuickPresetsCard.vue";
+import RecentJobsTable from "@/components/tasks/RecentJobsTable.vue";
 import RefreshToolbar from "@/components/RefreshToolbar.vue";
 import TaskCard from "@/components/tasks/TaskCard.vue";
 import { useAutoRefresh } from "@/composables/useAutoRefresh";
@@ -40,107 +43,11 @@ function navigateToNewTask() {
   router.push("/tasks/new");
 }
 
-const jobStatusColor = computed(() => {
-  return (status: string) => {
-    switch (status) {
-      case "running":
-        return "var(--accent-fg)";
-      case "completed":
-        return "var(--success-fg)";
-      case "failed":
-        return "var(--danger-fg)";
-      case "cancelled":
-        return "var(--text-tertiary)";
-      default:
-        return "var(--text-secondary)";
-    }
-  };
-});
-
 const hasActiveFilters = computed(
   () =>
     store.searchQuery.trim() !== "" || store.filterStatus !== "all" || store.filterType !== "all",
 );
 
-const stateLabel = computed(() => {
-  if (orchestrator.starting) return "Starting…";
-  switch (orchestrator.health?.health) {
-    case "healthy":
-      return "Running";
-    case "stale":
-      return "Stale";
-    case "stopped":
-      return "Stopped";
-    case "unknown":
-      return orchestrator.handle ? "Starting…" : "Unknown";
-    default:
-      return orchestrator.handle ? "Starting…" : "Idle";
-  }
-});
-
-const stateColorClass = computed(() => {
-  switch (orchestrator.health?.health) {
-    case "healthy":
-      return "state-color-healthy";
-    case "stale":
-      return "state-color-stale";
-    case "stopped":
-      return "state-color-stopped";
-    default:
-      return orchestrator.handle ? "state-color-starting" : "state-color-stopped";
-  }
-});
-
-// Reactive clock for uptime display (ticks every second)
-const nowMs = ref(Date.now());
-const _uptimeClock = setInterval(() => { nowMs.value = Date.now(); }, 1000);
-onUnmounted(() => clearInterval(_uptimeClock));
-
-const orchUptime = computed(() => {
-  if (!orchestrator.handle?.launchedAt) return null;
-  const launched = new Date(orchestrator.handle.launchedAt).getTime();
-  const diffSec = Math.floor((nowMs.value - launched) / 1000);
-  if (diffSec < 60) return `${diffSec}s`;
-  if (diffSec < 3600) return `${Math.floor(diffSec / 60)}m`;
-  const h = Math.floor(diffSec / 3600);
-  const m = Math.floor((diffSec % 3600) / 60);
-  return `${h}h ${m}m`;
-});
-
-const orchTaskProgress = computed(() => {
-  if (!store.stats) return null;
-  const { pending, inProgress, done, failed } = store.stats;
-  // Only show tasks in the active pipeline — queued (pending + in_progress).
-  // Done/failed are historical and shouldn't inflate the denominator.
-  const queued = pending + inProgress;
-  if (queued === 0) return null; // Nothing to process
-  return {
-    done,
-    failed,
-    pending,
-    inProgress,
-    total: queued,
-    pct: 0, // 0% because none of the queued tasks are done yet
-  };
-});
-
-function jobProgressPct(job: { tasksCompleted: number; taskCount: number }) {
-  if (job.taskCount === 0) return 0;
-  return Math.round((job.tasksCompleted / job.taskCount) * 100);
-}
-
-function jobProgressClass(status: string) {
-  switch (status) {
-    case "completed":
-      return "job-progress-success";
-    case "failed":
-      return "job-progress-danger";
-    case "running":
-      return "job-progress-accent";
-    default:
-      return "job-progress-neutral";
-  }
-}
 </script>
 
 <template>
@@ -262,145 +169,8 @@ function jobProgressClass(status: string) {
 
       <!-- Orchestrator Status Card + Quick Presets -->
       <div class="dashboard-cards">
-        <!-- Orchestrator Card -->
-        <div class="orch-card">
-          <div class="orch-card-header">
-            <div class="orch-card-title">
-              <span
-                class="orch-dot"
-                :class="orchestrator.isRunning ? 'dot-green' : orchestrator.isStale ? 'dot-warning' : 'dot-gray'"
-              />
-              Orchestrator
-              <span class="orch-state-label" :class="stateColorClass">
-                {{ stateLabel }}
-              </span>
-            </div>
-            <div class="orch-header-controls">
-              <select
-                v-if="orchestrator.isStopped && orchestrator.models.length > 0"
-                v-model="orchestrator.selectedModel"
-                class="orch-model-select"
-              >
-                <option v-for="m in orchestrator.models" :key="m.id" :value="m.id">
-                  {{ m.name }}
-                </option>
-              </select>
-              <button
-                v-if="orchestrator.isStopped"
-                class="orch-action-btn orch-start"
-                :disabled="orchestrator.starting"
-                @click="orchestrator.startOrchestrator()"
-              >
-                {{ orchestrator.starting ? "Starting…" : "Start" }}
-              </button>
-              <button
-                v-else
-                class="orch-action-btn orch-stop"
-                :disabled="orchestrator.stopping"
-                @click="orchestrator.stopOrchestrator()"
-              >
-                {{ orchestrator.stopping ? "Stopping…" : "Stop" }}
-              </button>
-            </div>
-          </div>
-          <div class="orch-card-stats">
-            <div class="orch-stat">
-              <span class="orch-stat-value">{{ orchestrator.health?.lastCycle ?? "—" }}</span>
-              <span class="orch-stat-label">Cycles</span>
-            </div>
-            <div class="orch-stat">
-              <span class="orch-stat-value">{{ orchestrator.health?.activeTasks?.length ?? 0 }}</span>
-              <span class="orch-stat-label">Active</span>
-            </div>
-            <div class="orch-stat">
-              <span
-                class="orch-stat-value"
-                :class="{ 'stale-text': orchestrator.isStale }"
-              >
-                {{ orchestrator.health?.heartbeatAgeSecs != null ? `${orchestrator.health.heartbeatAgeSecs}s` : "—" }}
-              </span>
-              <span class="orch-stat-label">Heartbeat</span>
-            </div>
-            <div v-if="orchUptime" class="orch-stat">
-              <span class="orch-stat-value">{{ orchUptime }}</span>
-              <span class="orch-stat-label">Uptime</span>
-            </div>
-          </div>
-          <!-- Task progress indicator -->
-          <div v-if="orchTaskProgress && orchestrator.isRunning" class="orch-progress-strip">
-            <div class="orch-progress-strip-header">
-              <span class="orch-progress-strip-label">Task Queue</span>
-              <span class="orch-progress-strip-value">
-                <template v-if="orchTaskProgress.inProgress > 0">⟳ {{ orchTaskProgress.inProgress }} running</template>
-                <template v-if="orchTaskProgress.inProgress > 0 && orchTaskProgress.pending > 0"> · </template>
-                <template v-if="orchTaskProgress.pending > 0">◌ {{ orchTaskProgress.pending }} pending</template>
-              </span>
-            </div>
-            <div class="orch-progress-bar">
-              <div
-                class="orch-progress-fill"
-                :style="{ width: orchTaskProgress.total > 0 ? `${Math.round((orchTaskProgress.inProgress / orchTaskProgress.total) * 100)}%` : '0%' }"
-              />
-            </div>
-          </div>
-          <div v-if="orchestrator.error" class="orch-error">{{ orchestrator.error }}</div>
-          <div class="orch-card-footer">
-            <button class="orch-monitor-link" @click="router.push('/tasks/monitor')">
-              Open Monitor →
-            </button>
-            <button
-              v-if="orchestrator.sessionUuid"
-              class="orch-monitor-link"
-              @click="router.push(`/session/${orchestrator.sessionUuid}/overview`)"
-            >
-              View Session →
-            </button>
-          </div>
-        </div>
-
-        <!-- Quick Presets Card -->
-        <div class="quick-presets-card">
-          <div class="quick-presets-header">
-            <span class="quick-presets-title">Quick Presets</span>
-            <button class="orch-monitor-link" @click="router.push('/tasks/presets')">
-              Manage →
-            </button>
-          </div>
-          <div v-if="presets.enabledPresets.length === 0" class="quick-presets-empty">
-            <svg width="28" height="28" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" style="opacity: 0.4; margin-bottom: 6px">
-              <rect x="3" y="3" width="18" height="18" rx="2" />
-              <path d="M12 8v8M8 12h8" />
-            </svg>
-            <span>No enabled presets</span>
-          </div>
-          <div v-else class="quick-preset-grid">
-            <button
-              v-for="preset in presets.enabledPresets.slice(0, 6)"
-              :key="preset.id"
-              class="quick-preset-card"
-              @click="router.push({ path: '/tasks/new', query: { presetId: preset.id } })"
-            >
-              <div class="quick-preset-icon">
-                <svg v-if="preset.taskType === 'session_summary'" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                  <path d="M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z" />
-                  <polyline points="14 2 14 8 20 8" />
-                  <line x1="16" y1="13" x2="8" y2="13" /><line x1="16" y1="17" x2="8" y2="17" /><polyline points="10 9 9 9 8 9" />
-                </svg>
-                <svg v-else-if="preset.taskType === 'daily_digest' || preset.taskType === 'weekly_digest'" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                  <rect x="3" y="4" width="18" height="18" rx="2" ry="2" /><line x1="16" y1="2" x2="16" y2="6" /><line x1="8" y1="2" x2="8" y2="6" /><line x1="3" y1="10" x2="21" y2="10" />
-                </svg>
-                <svg v-else width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                  <polygon points="12 2 15.09 8.26 22 9.27 17 14.14 18.18 21.02 12 17.77 5.82 21.02 7 14.14 2 9.27 8.91 8.26 12 2" />
-                </svg>
-              </div>
-              <div class="quick-preset-info">
-                <span class="quick-preset-name">{{ preset.name }}</span>
-                <span v-if="preset.builtin" class="quick-preset-builtin">Built-in</span>
-              </div>
-              <span class="quick-preset-arrow">→</span>
-            </button>
-          </div>
-        </div>
+        <OrchestratorStatusCard :stats="store.stats" />
+        <QuickPresetsCard :presets="presets.enabledPresets" />
       </div>
 
       <!-- Filter Row -->
@@ -508,63 +278,10 @@ function jobProgressClass(status: string) {
       </div>
 
       <!-- Recent Jobs Section -->
-      <div v-if="!store.loading && !store.error && store.jobs.length > 0" class="jobs-section">
-        <h2 class="jobs-section__title">Recent Jobs</h2>
-        <div class="jobs-table-wrap">
-          <table class="jobs-table">
-            <thead>
-              <tr>
-                <th>Name</th>
-                <th>Status</th>
-                <th>Progress</th>
-                <th>Created</th>
-              </tr>
-            </thead>
-            <tbody>
-              <tr v-for="job in store.jobs" :key="job.id">
-                <td class="jobs-table__name">{{ job.name }}</td>
-                <td>
-                  <span class="job-status-badge-wrap">
-                    <span
-                      class="job-status-dot"
-                      :style="{ background: jobStatusColor(job.status) }"
-                    />
-                    <span
-                      class="job-status-badge"
-                      :style="{ color: jobStatusColor(job.status) }"
-                    >
-                      {{ job.status }}
-                    </span>
-                  </span>
-                </td>
-                <td class="jobs-table__progress">
-                  <div class="job-progress-bar-wrap">
-                    <div class="job-progress-bar">
-                      <div
-                        class="job-progress-fill"
-                        :class="jobProgressClass(job.status)"
-                        :style="{
-                          width: `${jobProgressPct(job)}%`,
-                        }"
-                      />
-                    </div>
-                    <span class="job-progress-pct">{{ jobProgressPct(job) }}%</span>
-                  </div>
-                  <span class="jobs-table__counts">
-                    {{ job.tasksCompleted }}/{{ job.taskCount }} done
-                  </span>
-                  <span v-if="job.tasksFailed > 0" class="jobs-table__failed">
-                    · {{ job.tasksFailed }} failed
-                  </span>
-                </td>
-                <td class="jobs-table__date">
-                  {{ formatDate(job.createdAt) }}
-                </td>
-              </tr>
-            </tbody>
-          </table>
-        </div>
-      </div>
+      <RecentJobsTable
+        v-if="!store.loading && !store.error && store.jobs.length > 0"
+        :jobs="store.jobs"
+      />
     </div>
   </div>
 </template>
@@ -705,340 +422,6 @@ function jobProgressClass(status: string) {
   margin-top: 16px;
 }
 
-.orch-card,
-.quick-presets-card {
-  background: var(--canvas-subtle);
-  border: 1px solid var(--border-default);
-  border-radius: var(--radius-lg);
-  padding: 16px 20px;
-  min-width: 0;
-  overflow: hidden;
-}
-
-.orch-card-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 14px;
-}
-
-.orch-card-title {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-size: 0.875rem;
-  font-weight: 600;
-  color: var(--text-primary);
-}
-
-.orch-dot {
-  width: 8px;
-  height: 8px;
-  border-radius: 50%;
-  flex-shrink: 0;
-}
-
-.dot-green {
-  background: var(--success-fg);
-  box-shadow: 0 0 6px var(--success-muted);
-}
-
-.dot-gray {
-  background: var(--neutral-emphasis);
-}
-
-.dot-warning {
-  background: var(--warning-fg);
-  box-shadow: 0 0 6px var(--warning-muted);
-}
-
-.orch-state-label {
-  font-size: 0.6875rem;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.04em;
-}
-
-.orch-action-btn {
-  padding: 4px 14px;
-  font-size: 0.6875rem;
-  font-weight: 600;
-  border: none;
-  border-radius: var(--radius-md);
-  cursor: pointer;
-  transition: background var(--transition-fast);
-}
-
-.orch-header-controls {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.orch-model-select {
-  padding: 3px 8px;
-  font-size: 0.6875rem;
-  background: var(--surface-tertiary);
-  color: var(--text-secondary);
-  border: 1px solid var(--border-default);
-  border-radius: var(--radius-md);
-  cursor: pointer;
-  max-width: 140px;
-}
-
-.orch-model-select:focus {
-  border-color: var(--accent-fg);
-  outline: none;
-}
-
-.orch-action-btn:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
-.orch-start {
-  background: var(--success-fg);
-  color: var(--text-inverse);
-}
-
-.orch-start:hover:not(:disabled) {
-  background: var(--success-emphasis);
-}
-
-.orch-stop {
-  background: var(--danger-fg);
-  color: var(--text-inverse);
-}
-
-.orch-stop:hover:not(:disabled) {
-  background: var(--danger-emphasis);
-}
-
-.orch-card-stats {
-  display: flex;
-  gap: 24px;
-  margin-bottom: 12px;
-}
-
-.orch-stat {
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-}
-
-.orch-stat-value {
-  font-size: 1.125rem;
-  font-weight: 700;
-  color: var(--text-primary);
-  font-variant-numeric: tabular-nums;
-}
-
-.stale-text {
-  color: var(--warning-fg);
-}
-
-.state-color-healthy {
-  color: var(--success-fg);
-}
-.state-color-stale {
-  color: var(--warning-fg);
-}
-.state-color-stopped {
-  color: var(--neutral-emphasis);
-}
-.state-color-starting {
-  color: var(--accent-fg);
-}
-
-.job-progress-success {
-  background: var(--success-fg);
-}
-.job-progress-danger {
-  background: var(--danger-fg);
-}
-.job-progress-accent {
-  background: var(--accent-fg);
-}
-.job-progress-neutral {
-  background: var(--text-tertiary);
-}
-
-.orch-stat-label {
-  font-size: 0.6875rem;
-  color: var(--text-tertiary);
-  text-transform: uppercase;
-  letter-spacing: 0.04em;
-}
-
-.orch-error {
-  font-size: 0.75rem;
-  color: var(--danger-fg);
-  margin-bottom: 8px;
-}
-
-.orch-card-footer {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.orch-progress-strip {
-  margin-bottom: 12px;
-}
-
-.orch-progress-strip-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 4px;
-}
-
-.orch-progress-strip-label {
-  font-size: 0.6875rem;
-  color: var(--text-tertiary);
-}
-
-.orch-progress-strip-value {
-  font-size: 0.6875rem;
-  color: var(--accent-fg);
-  font-weight: 600;
-  font-variant-numeric: tabular-nums;
-}
-
-.orch-progress-bar {
-  height: 4px;
-  background: var(--border-default);
-  border-radius: 999px;
-  overflow: hidden;
-}
-
-.orch-progress-fill {
-  height: 100%;
-  border-radius: 999px;
-  background: var(--accent-fg);
-  transition: width 0.8s cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-.orch-monitor-link {
-  background: none;
-  border: none;
-  color: var(--accent-fg);
-  font-size: 0.75rem;
-  font-weight: 500;
-  cursor: pointer;
-  padding: 0;
-  transition: opacity var(--transition-fast);
-}
-
-.orch-monitor-link:hover {
-  opacity: 0.8;
-}
-
-/* Quick Presets */
-.quick-presets-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 14px;
-}
-
-.quick-presets-title {
-  font-size: 0.875rem;
-  font-weight: 600;
-  color: var(--text-primary);
-}
-
-.quick-presets-empty {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 2px;
-  font-size: 0.8125rem;
-  color: var(--text-placeholder);
-  text-align: center;
-  padding: 24px 0;
-}
-
-.quick-preset-grid {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-
-.quick-preset-card {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  padding: 10px 12px;
-  background: var(--canvas-default);
-  border: 1px solid var(--border-subtle);
-  border-radius: var(--radius-md);
-  cursor: pointer;
-  transition:
-    border-color var(--transition-fast),
-    background var(--transition-fast);
-  text-align: left;
-  color: inherit;
-  font: inherit;
-}
-
-.quick-preset-card:hover {
-  border-color: var(--accent-fg);
-  background: var(--accent-muted);
-}
-
-.quick-preset-icon {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  border-radius: var(--radius-md);
-  background: var(--accent-muted);
-  color: var(--accent-fg);
-  flex-shrink: 0;
-}
-
-.quick-preset-info {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  flex: 1;
-  min-width: 0;
-}
-
-.quick-preset-name {
-  font-size: 0.8125rem;
-  font-weight: 600;
-  color: var(--text-primary);
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.quick-preset-builtin {
-  font-size: 0.5625rem;
-  padding: 1px 5px;
-  border-radius: var(--radius-sm, 4px);
-  background: var(--canvas-subtle);
-  color: var(--text-tertiary);
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.03em;
-  flex-shrink: 0;
-}
-
-.quick-preset-arrow {
-  color: var(--text-tertiary);
-  font-size: 0.875rem;
-  flex-shrink: 0;
-  transition: transform var(--transition-fast);
-}
-
-.quick-preset-card:hover .quick-preset-arrow {
-  transform: translateX(2px);
-  color: var(--accent-fg);
-}
 
 /* ── Filter Row ──────────────────────────────────────────── */
 .filter-row {
@@ -1209,133 +592,6 @@ function jobProgressClass(status: string) {
   gap: 8px;
 }
 
-/* ── Recent Jobs Section ─────────────────────────────────── */
-.jobs-section {
-  margin-top: 32px;
-  padding-top: 24px;
-  border-top: 1px solid var(--border-default);
-}
-
-.jobs-section__title {
-  font-size: 1rem;
-  font-weight: 600;
-  color: var(--text-primary);
-  margin: 0 0 14px;
-  letter-spacing: -0.01em;
-}
-
-.jobs-table-wrap {
-  overflow-x: auto;
-  border: 1px solid var(--border-default);
-  border-radius: 8px;
-  background: var(--canvas-subtle);
-}
-
-.jobs-table {
-  width: 100%;
-  border-collapse: collapse;
-  font-size: 0.8125rem;
-}
-
-.jobs-table th {
-  text-align: left;
-  padding: 10px 14px;
-  font-size: 0.6875rem;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.04em;
-  color: var(--text-tertiary);
-  border-bottom: 1px solid var(--border-default);
-  white-space: nowrap;
-}
-
-.jobs-table td {
-  padding: 10px 14px;
-  color: var(--text-secondary);
-  border-bottom: 1px solid
-    var(--border-muted);
-  vertical-align: middle;
-}
-
-.jobs-table tbody tr:last-child td {
-  border-bottom: none;
-}
-
-.jobs-table tbody tr:hover {
-  background: var(--neutral-subtle);
-}
-
-.jobs-table__name {
-  color: var(--text-primary);
-  font-weight: 500;
-}
-
-.job-status-badge {
-  font-size: 0.75rem;
-  font-weight: 600;
-  text-transform: capitalize;
-}
-
-.jobs-table__progress {
-  font-variant-numeric: tabular-nums;
-}
-
-.jobs-table__counts {
-  color: var(--text-secondary);
-}
-
-.jobs-table__failed {
-  color: var(--danger-fg);
-  font-weight: 500;
-}
-
-.jobs-table__date {
-  color: var(--text-tertiary);
-  white-space: nowrap;
-}
-
-.job-status-badge-wrap {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-}
-
-.job-status-dot {
-  width: 7px;
-  height: 7px;
-  border-radius: 50%;
-  flex-shrink: 0;
-}
-
-.job-progress-bar-wrap {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  margin-bottom: 4px;
-}
-
-.job-progress-bar {
-  flex: 1;
-  height: 3px;
-  background: var(--border-default);
-  border-radius: 999px;
-  overflow: hidden;
-  min-width: 60px;
-}
-
-.job-progress-fill {
-  height: 100%;
-  border-radius: 999px;
-  transition: width 0.6s ease;
-}
-
-.job-progress-pct {
-  font-size: 0.625rem;
-  color: var(--text-tertiary);
-  font-variant-numeric: tabular-nums;
-  min-width: 28px;
-  text-align: right;
-}
 
 /* ── Utilities ───────────────────────────────────────────── */
 .spin-animation {
